requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
Django==4.1.3
django-environ==0.9.0
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
PyJWT==2.6.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
zipp==3.10.0
